Cleanup, fix UI, add flags
