+++ src/assets/avatars/avatars.js
@@ -1,3 +1,5 @@
+import { getReticulumFetchUrl } from "../../utils/phoenix-utils";
+
 export const avatars = [
   {
     id: "botdefault",
@@ -40,3 +42,26 @@ export const avatars = [
     model: "https://asset-bundles-prod.reticulum.io/bots/BotWoody_Avatar-0140485a23.gltf"
   }
 ];
+
+export const AVATAR_TYPES = {
+  LEGACY: "legacy",
+  SKINNABLE: "skinnable",
+  URL: "url"
+};
+
+const legacyAvatarIds = avatars.map(a => a.id);
+export function getAvatarType(avatarId) {
+  if (!avatarId || legacyAvatarIds.indexOf(avatarId) !== -1) return AVATAR_TYPES.LEGACY;
+  if (avatarId.startsWith("http")) return AVATAR_TYPES.URL;
+  return AVATAR_TYPES.SKINNABLE;
+}
+
+export function getAvatarSrc(avatarId, cacheVersion = 0) {
+  switch (getAvatarType(avatarId)) {
+    case AVATAR_TYPES.LEGACY:
+      return `#${avatarId}`;
+    case AVATAR_TYPES.SKINNABLE:
+      return getReticulumFetchUrl(`/api/v1/avatars/${avatarId}/avatar.gltf?v=${cacheVersion}`);
+  }
+  return avatarId;
+}
